# Conflicts:
#	include/darmok/protobuf.hpp
#	include/darmok/scene_serialize.hpp

+++ src/asset_core.cpp
@@ -1181,7 +1181,7 @@ namespace darmok
 
         try
         {
-            cli.parse(args.size(), args.data());
+            cli.parse(static_cast<int>(args.size()), args.data());
             cfg.fix(cli);
 
             if (cfg.dry)
